requirements.txt
gym
keras
keras-rl2
numpy==1.18.5
requests
scikit-learn
scipy
sklearn
tensorflow
chainer
chainerrl
libinjection-python
waf-brain
sqlparse